split postprocessingservice into root, wb and axl
