+++ src/CSharp App/Services/PostProcessingService.cs
@@ -14,7 +14,7 @@ using YamlDotNet.Serialization.NamingConventions;
 
 namespace VolumetricSelection2077.Services;
 
-public class PostProcessingService
+public partial class PostProcessingService
 {
     private readonly Progress _progress;
     private readonly SettingsService _settingsService;
@@ -35,7 +35,7 @@ public class PostProcessingService
     {
         _progress.AddTarget(2, Progress.ProgressSections.Finalization);
 
-        var nullCheckedSectors = rawSectors.Where(s => s != null).ToList();
+        var nullCheckedSectors = rawSectors.Where(s => s != null).ToList() as List<AxlRemovalSector>;
         if (nullCheckedSectors.Count == 0)
         {
             Logger.Warning("No sectors intersect, no output file generated!");
@@ -64,363 +64,6 @@ public class PostProcessingService
         }
         
         _progress.AddCurrent(1, Progress.ProgressSections.Finalization);
-    }
-
-    /// <summary>
-    /// Saves the AxlRemovalFile as an xl file in either JSON or YAML depending on the state of the SettingsService
-    /// </summary>
-    /// <param name="axlRemovalFile"></param>
-    private void SaveAsRemoval(AxlRemovalFile axlRemovalFile)
-    {
-        var nodeCount = axlRemovalFile.Streaming.Sectors.Select(s => s?.NodeDeletions.Count).Sum();
-        var actorCount = axlRemovalFile.Streaming.Sectors.SelectMany(s => s?.NodeDeletions ?? new()).Sum(n => n.ActorDeletions?.Count);
-        Logger.Success($"Found {nodeCount} node deletions and {actorCount} actor deletions across {axlRemovalFile.Streaming.Sectors.Count} sectors");
-        
-        string outputFilePath;
-        if (_settingsService.SaveToArchiveMods)
-            outputFilePath = Path.Join(_settingsService.GameDirectory, "archive", "pc", "mod", _settingsService.OutputFilename) + ".xl";
-        else
-            outputFilePath = Path.Join(_settingsService.OutputDirectory, _settingsService.OutputFilename) + ".xl";
-        
-        var (outputContent, mergeChanges) = SerializeAxlRemovalFile(axlRemovalFile, outputFilePath);
-        
-        Directory.CreateDirectory(Path.GetDirectoryName(outputFilePath));
-        
-        if (!File.Exists(outputFilePath))
-        {
-            if (_settingsService.SaveMode == SaveFileMode.Subtract)
-            {
-                Logger.Error($"No Existing File to remove from found at {outputFilePath}");
-                return;
-            }
-            
-            File.WriteAllText(outputFilePath, outputContent);
-            Logger.Info($"Created file {outputFilePath}");
-            WriteBackupFile(outputFilePath, outputContent);
-            return;
-        }
-
-        switch (_settingsService.SaveMode)
-        {
-            case SaveFileMode.Extend:
-            {
-                File.WriteAllText(outputFilePath, outputContent);
-                var newSectorS = mergeChanges?.newSectors != 1 ? "s" : "";
-                var newNodesS = mergeChanges?.newNodes != 1 ? "s" : "";
-                var newActorsS = mergeChanges?.newActors != 1 ? "s" : "";
-                Logger.Info($"Extended file {outputFilePath} with {mergeChanges.newSectors} new sector{newSectorS}, {mergeChanges.newNodes} new node{newNodesS}, {mergeChanges.newActors} new actor{newActorsS}.");
-                WriteBackupFile(outputFilePath, outputContent);
-                return;
-            }
-            case SaveFileMode.Overwrite:
-                File.WriteAllText(outputFilePath, outputContent);
-                Logger.Info($"Overwrote file {outputFilePath}");
-                WriteBackupFile(outputFilePath, outputContent);
-                return;
-            case SaveFileMode.New:
-                var newOutputFilePath = GetOutputFilename(outputFilePath);
-                File.WriteAllText(newOutputFilePath, outputContent);
-                Logger.Info($"Created file {newOutputFilePath}");
-                WriteBackupFile(newOutputFilePath, outputContent);
-                return;
-            case SaveFileMode.Subtract:
-                File.WriteAllText(outputFilePath, outputContent);
-                var remSectorS = mergeChanges?.newSectors != 1 ? "s" : "";
-                var remNodesS = mergeChanges?.newNodes != 1 ? "s" : "";
-                var remActorsS = mergeChanges?.newActors != 1 ? "s" : "";
-                Logger.Info($"Removed {mergeChanges.newSectors} sector{remSectorS}, {mergeChanges.newNodes} node{remNodesS}, {mergeChanges.newActors} actor{remActorsS} from file {outputFilePath}.");
-                WriteBackupFile(outputFilePath, outputContent);
-                return;
-            default:
-                throw new ArgumentOutOfRangeException();
-        }
-    }
-    
-    private (string, MergeChanges) SerializeAxlRemovalFile(AxlRemovalFile axlRemovalFile, string outputFilePath)
-    {
-        var mergeChanges = new MergeChanges();
-
-        if (File.Exists(outputFilePath))
-        {
-            AxlRemovalFile? existingRemovals = null;
-
-            if (_settingsService.SaveMode is SaveFileMode.Extend or SaveFileMode.Subtract)
-            {
-                var fileContent = File.ReadAllText(outputFilePath);
-                existingRemovals = UtilService.TryParseAxlRemovalFile(fileContent);
-                if (existingRemovals == null)
-                    throw new FileLoadException($"Failed to find or parse existing Removal File at {outputFilePath}!"); 
-            }
-
-            (axlRemovalFile, mergeChanges) = _settingsService.SaveMode switch
-            {
-                SaveFileMode.Extend => MergeSectors(existingRemovals!, axlRemovalFile),
-                SaveFileMode.Subtract => SubtractRemovals(existingRemovals!, axlRemovalFile),
-                _ => (axlRemovalFile, mergeChanges)
-            };
-        }
-        
-        string outputContent;
-
-        switch (_settingsService.SaveFileFormat)
-        {
-            case  SaveFileFormat.ArchiveXLJson:
-                outputContent = JsonConvert.SerializeObject(axlRemovalFile, JsonSerializerPresets.WorldBuilder);
-                break;
-            case  SaveFileFormat.ArchiveXLYaml:
-                var serializer = new SerializerBuilder()
-                    .WithNamingConvention(CamelCaseNamingConvention.Instance)
-                    .ConfigureDefaultValuesHandling(DefaultValuesHandling.OmitNull)
-                    .Build();
-
-                outputContent = serializer.Serialize(axlRemovalFile);
-                break;
-            default:
-                throw new ArgumentOutOfRangeException();
-        }
-        
-        return (outputContent, mergeChanges);
-    }
-    
-    /// <summary>
-    /// Merge Two AxlRemovalFiles
-    /// </summary>
-    /// <param name="removals1"></param>
-    /// <param name="removals2"></param>
-    /// <returns></returns>
-    private static (AxlRemovalFile, MergeChanges) MergeSectors(AxlRemovalFile removals1, AxlRemovalFile removals2)
-    {
-        var changeCount = new MergeChanges();
-        var newSectors = removals2.Streaming.Sectors;
-        var oldSectors = removals1.Streaming.Sectors;
-        
-        Dictionary<string, AxlRemovalSector> mergedDict = oldSectors.ToDictionary(x => x.Path);
-
-        foreach (var newSector in newSectors)
-        {
-            if (mergedDict.TryGetValue(newSector.Path, out AxlRemovalSector existingSector))
-            {
-                Dictionary<int, AxlRemovalNodeDeletion> mergedNodes =
-                    existingSector.NodeDeletions.ToDictionary(x => x.Index);
-                foreach (var newNode in newSector.NodeDeletions)
-                {
-                    if (mergedNodes.TryGetValue(newNode.Index, out AxlRemovalNodeDeletion existingNode))
-                    {
-                        if (newNode.ActorDeletions != null || 
-                            newNode.ActorDeletions?.Count > 0 ||
-                            existingNode.ActorDeletions?.Count != null ||
-                            existingNode.ActorDeletions?.Count > 0)
-                        {
-                            existingNode.ExpectedActors =  newNode.ExpectedActors ?? existingNode.ExpectedActors;
-                            HashSet<int> actorSet = new HashSet<int>(newNode.ActorDeletions ?? new List<int>());
-                            actorSet.UnionWith(existingNode.ActorDeletions ?? new List<int>());
-                            existingNode.ActorDeletions = actorSet.ToList();
-                            changeCount.newActors += actorSet.Count - existingNode.ActorDeletions.Count;
-                        }
-                    }
-                    else
-                    {
-                        mergedNodes[newNode.Index] = newNode;
-                        changeCount.newNodes++;
-                        changeCount.newActors += newNode.ActorDeletions?.Count ?? 0;
-                    }
-                }
-                existingSector.NodeDeletions = mergedNodes.Values.ToList();
-            }
-            else
-            {
-                mergedDict[newSector.Path] = newSector;
-                changeCount.newSectors++;
-                changeCount.newNodes += newSector.NodeDeletions?.Count ?? 0;
-                foreach (var newNode in newSector.NodeDeletions)
-                    changeCount.newActors += newNode.ActorDeletions?.Count ?? 0;
-            }
-        }
-        var mergedSectors = mergedDict.Values.ToList();
-        var mergedRemovalFile = new AxlRemovalFile
-        {
-            Streaming = new AxlRemovalStreaming
-            {
-                Sectors = mergedSectors,
-            }
-        };
-        return (mergedRemovalFile, changeCount);
-    }
-
-    /// <summary>
-    /// Subtracts the AxlRemovalFile from the baseFile
-    /// </summary>
-    /// <param name="baseFile"></param>
-    /// <param name="subtraction"></param>
-    /// <returns></returns>
-    private static (AxlRemovalFile, MergeChanges) SubtractRemovals(AxlRemovalFile baseFile, AxlRemovalFile subtraction)
-    {
-        var mc = new MergeChanges();
-        foreach (var sector in subtraction.Streaming.Sectors)
-        {
-            var baseSector = baseFile.Streaming.Sectors.FirstOrDefault(s => s.Path == sector.Path);
-            if (baseSector == null)
-                continue;
-            
-            foreach (var node in sector.NodeDeletions)
-            {
-                var baseNode = baseSector.NodeDeletions.FirstOrDefault(n => n.Index == node.Index);
-                if (baseNode == null)
-                    continue;
-                
-                if (node.ActorDeletions != null && baseNode.ActorDeletions != null)
-                {
-                    var baseActorSet = new HashSet<int>(baseNode.ActorDeletions);
-                    var subActorSet = new HashSet<int>(node.ActorDeletions);
-                    var diff = baseActorSet.Except(subActorSet).ToList();
-                    baseNode.ActorDeletions = diff;
-                    mc.newActors += diff.Count;
-                    
-                    if (baseNode.ActorDeletions.Count == 0)
-                    {
-                        baseSector.NodeDeletions.Remove(baseNode);
-                        mc.newNodes++;
-                    }
-                }
-                else
-                {
-                    baseSector.NodeDeletions.Remove(baseNode);
-                    mc.newNodes++;
-                }
-            }
-
-            if (baseSector.NodeDeletions.Count == 0)
-            {
-                baseFile.Streaming.Sectors.Remove(baseSector);
-                mc.newSectors++;
-            }
-        }
-        
-        return (baseFile, mc);
-    }
-    
-    /// <summary>
-    /// Saves the AxlRemovalFile as a world builder prefab
-    /// </summary>
-    /// <param name="axlRemovalFile"></param>
-    private void SaveAsPrefab(AxlRemovalFile axlRemovalFile)
-    {
-        var favoritesPath = _settingsService.SaveToArchiveMods ?
-            Path.Join(_settingsService.GameDirectory, "bin", "x64", "plugins", "cyber_engine_tweaks", "mods", "entSpawner", "data", "favorite", "GeneratedByVS2077.json") :
-            Path.Join(_settingsService.OutputDirectory, "GeneratedByVS2077.json");
-        var logMessage = "";
-        
-        
-        var favRoot = new FavoritesRoot
-        {
-            Name = "GeneratedByVS2077",
-            Favorites = new()
-        };
-
-        var convertedData = _removalToWorldBuilderConverter.Convert(axlRemovalFile, _settingsService.OutputFilename);
-        Logger.Success($"Found {convertedData.Children.OfType<PositionableGroup>().Sum(g => g.Children.Count)} WorldBuilder elements");
-        if (!File.Exists(favoritesPath))
-        {
-            if (_settingsService.SaveMode == SaveFileMode.Subtract)
-            {
-                Logger.Error($"No Existing File to remove from found at {favoritesPath}");
-                return;
-            }
-            
-            favRoot.Favorites.Add(new Favorite
-            {
-                Name =  _settingsService.OutputFilename,
-                Data = (Positionable)convertedData
-            });
-            logMessage = $"Created prefab {_settingsService.OutputFilename} at {favoritesPath}";
-
-        }
-        else
-        {
-            var existingFavorites = JsonConvert.DeserializeObject<FavoritesRoot>(File.ReadAllText(favoritesPath), JsonSerializerPresets.WorldBuilder);
-            
-            switch (_settingsService.SaveMode)
-            {
-                case SaveFileMode.Overwrite:
-                    var existingOverwritePrefab = existingFavorites?.Favorites.FirstOrDefault(f => f.Name == _settingsService.OutputFilename);
-                    if (existingOverwritePrefab != null)
-                    {
-                        existingOverwritePrefab.Data =
-                            (Positionable)convertedData;
-                        logMessage = $"Overwrote prefab {_settingsService.OutputFilename} at {favoritesPath}";
-                    }
-                    else
-                        goto newPrefab;
-                    break;
-                case SaveFileMode.Extend:
-                    var existingExtendPrefab = existingFavorites?.Favorites.FirstOrDefault(f => f.Name == _settingsService.OutputFilename);
-                    if (existingExtendPrefab != null)
-                    {
-                        var existingElementsCount = existingExtendPrefab.Data.Children.OfType<PositionableGroup>().Sum(g => g.Children.Count);
-                        existingExtendPrefab.Data = WorldBuilderMergingService.Merge(existingExtendPrefab, new Favorite
-                        {
-                            Name = _settingsService.OutputFilename,
-                            Data = (Positionable)convertedData
-                        }).Data;
-                        var newElementsCount = existingExtendPrefab.Data.Children.OfType<PositionableGroup>().Sum(g => g.Children.Count);
-                        logMessage = $"Extended prefab {_settingsService.OutputFilename} with {newElementsCount - existingElementsCount} new elements at {favoritesPath}";
-                    }
-                    else
-                        goto newPrefab;
-                    break;
-                case SaveFileMode.New:
-                    newPrefab:
-                    var existingNewPrefab = existingFavorites?.Favorites.FirstOrDefault(f => f.Name == _settingsService.OutputFilename);
-                    if (existingNewPrefab == null)
-                    {
-                        existingFavorites?.Favorites.Add(new Favorite
-                        {
-                            Name = _settingsService.OutputFilename,
-                            Data = (Positionable)convertedData
-                        });
-                        logMessage = $"Created prefab {_settingsService.OutputFilename} at {favoritesPath}";
-                    }
-                    else
-                    {
-                        var newCount = existingFavorites?.Favorites.Count(f => f.Name.StartsWith(_settingsService.OutputFilename));
-                        var newOutputFilename = $"{_settingsService.OutputFilename}+{newCount}";
-                        existingFavorites?.Favorites.Add(new Favorite
-                        {
-                            Name = newOutputFilename,
-                            Data = (Positionable)convertedData
-                        });
-                        logMessage = $"Created prefab {newOutputFilename} at {favoritesPath}";
-                    }
-                    break;
-                case SaveFileMode.Subtract:
-                    var existingSubtractPrefab = existingFavorites?.Favorites.FirstOrDefault(f => f.Name == _settingsService.OutputFilename);
-                    if (existingSubtractPrefab != null)
-                    {
-                        var existingElementsCount = existingSubtractPrefab.Data.Children.OfType<PositionableGroup>().Sum(g => g.Children.Count);
-                        existingSubtractPrefab.Data = WorldBuilderMergingService.Subtract(existingSubtractPrefab, new Favorite
-                        {
-                            Name = _settingsService.OutputFilename,
-                            Data = (Positionable)convertedData
-                        }).Data;
-                        var newElementsCount = existingSubtractPrefab.Data.Children.OfType<PositionableGroup>().Sum(g => g.Children.Count);
-                        logMessage = $"Removed {(newElementsCount - existingElementsCount) * -1} elements from prefab {_settingsService.OutputFilename} at {favoritesPath}";
-                    }
-                    else
-                    {
-                        Logger.Error($"No Existing prefab to remove from found at {favoritesPath}");
-                        return;
-                    }
-                    break;
-            }
-            
-            if (existingFavorites != null)
-                favRoot.Favorites.AddRange(existingFavorites.Favorites);
-        }
-
-        var serialized = JsonConvert.SerializeObject(favRoot, JsonSerializerPresets.WorldBuilder);
-        File.WriteAllText(favoritesPath, serialized);
-        Logger.Info(logMessage);
-        WriteBackupFile(favoritesPath, serialized);
     }
     
     private static string GetOutputFilename(string outputFilename)
